Refactored name to ManageNotificationResultService
